docs: README 2026-02-22 반영 및 DB/모델 변경 문서화
- README: 백엔드/DB 안정화(update-select 2단계, infusion_rate float/NUMERIC, view_token_expires_fix), Traceback 로깅, search_error.py, 문서 경로
- backend/models: IVRecord.infusion_rate float (이미 적용분 커밋)
- 마이그레이션: iv_records NUMERIC, view_station_dashboard + token_expires_at
- PROMPT_MEAL_SYNC_SUBMODAL_STALE: 식단 일괄 생성 후 서브모달 미갱신 디버깅 프롬프트

+++ backend/models.py
@@ -43,7 +43,7 @@ class IVRecord(BaseModel):
     id: Optional[int] = None
     admission_id: str
     photo_url: Optional[str] = None
-    infusion_rate: Optional[int] = None
+    infusion_rate: Optional[float] = None
     created_at: Optional[datetime] = None
 
 class MealRequest(BaseModel):
@@ -94,7 +94,7 @@ class VitalSignCreate(BaseModel):
 class IVRecordCreate(BaseModel):
     admission_id: str
     photo_url: Optional[str] = None
-    infusion_rate: int  # cc/hr
+    infusion_rate: float  # cc/hr
 
 class MealRequestCreate(BaseModel):
     admission_id: str
